% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2RangesList.R
\name{df2RangesList}
\alias{df2RangesList}
\title{Transform annotated peptide data to list of IRanges}
\usage{
df2RangesList(df, n = "name")
}
\arguments{
\item{df}{A peptide-level group comparision result (data.frame) annotated by posprot::annotate_gcr()}

\item{n}{The name of RangesList object returned by this function}
}
\value{
This function returns a RangesList object
}
\description{
Transform annotated peptide data to list of IRanges
}
