% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_neoCN.R
\name{find_neoCN_pairs}
\alias{find_neoCN_pairs}
\title{Find paired neo N- and neo C-terminal peptides over a list of proteins}
\usage{
find_neoCN_pairs(l)
}
\arguments{
\item{l}{is a list of IRanges representing peptides matched to proteins}
}
\value{
This function returns a list of IRanges representing all paired peptides found in the input data.
}
\description{

}
\details{
---XXXXX.XXXXX--- ==> ---XXXXXXXXXX---
}
