% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.comparision.labels.R
\name{get.comparison.labels}
\alias{get.comparison.labels}
\title{Extract comparision labels from a MSstats group comparision result}
\usage{
get.comparison.labels(x)
}
\arguments{
\item{x}{MSstats group comparision result as generated by \code{\link[MSstats]{groupComparison}}}
}
\value{
This function returns a character vector containing all available labels.
}
\description{
Accessor function for the ComparisonResult$Label component of objects returned by \code{\link[MSstats]{groupComparison}}
}
\details{
\code{\link[MSstats]{groupComparison}} returns a complex, but primitive object. Accordingly, there are not accessor function available. This function allows controlled retrivial.
}
\examples{
get.comparison.labels(sample.data)
}
