% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.MSstats_gcr.R
\name{is.MSstats_gcr}
\alias{is.MSstats_gcr}
\title{Function to check if an object was generated by MSstats::groupComparison() function.}
\usage{
is.MSstats_gcr(x)
}
\arguments{
\item{x}{any R object}
}
\value{
is.MSstats_gcr returns TRUE if its argument was most likely generated by groupComparison()
and FALSE otherwise.
}
\description{
MSstats::groupComparison() does not return propper S3 or S4 objects. This function can be used to verifiy if an object was generated by groupComparision().
}
\details{
This function checks for the expected object structure as described for MStats version 3.7.3. The expected column names for the ComparisonResults component are exported to the environment variable exp.names.
}
\examples{
is.MSstats_gcr(sample.data)
}
\seealso{
\code{\link[MSstats]{groupComparison}}
}
