% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_neoCN.R
\name{neoCN}
\alias{neoCN}
\title{Find paired neo C-terminal neo N-terminal peptides}
\usage{
neoCN(x, united = TRUE)
}
\arguments{
\item{x}{IRanges object}

\item{united}{Pairs object}
}
\value{
Function returns an IRanges object by default. If united is set to FALSE a pairs object
returned.
}
\description{
This function looks for adjacent peptides (peptides being positioned head-to-tail at zero distance)
taking the specificity of the test protease and the working protease into account.
The function also checks for the presence of N-terminal labels on the peptides. These indicate that a peptide lies downstream
of a cleavage site. Pairs of adjacent peptides, passing all specificity checks, are unified
and returned as a single IRange.
}
\details{
---XXXXX.XXXXX--- ==> ---XXXXXXXXXX---

This function expects test.prot.spec and work.prot.spec to be present in the environment.
}
