% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{sample.data}
\alias{sample.data}
\title{A sample data set generated using MSstats peptide-level inference}
\format{A complicated R object of type list.}
\source{
Generated using the R script found in data-raw
}
\usage{
is.MSstats_gcr(sample.data)
}
\description{
A sample data set generated using MSstats peptide-level inference
}
\keyword{datasets}
