% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificity.R
\name{spec}
\alias{spec}
\title{Retrieve protease specificity by name}
\usage{
spec(protease = "Trypsin")
}
\arguments{
\item{protease}{Name (character) of the protease}
}
\value{
A character vector containing amino acids cleaved by the protease
}
\description{
Retrieve protease specificity by name
}
\examples{
spec("GluC")
}
